move coldtype.animation namespace to coldtype.time to avoid clashing with @animation decorator also available under coldtype
