+++ coldtype/time/__init__.py
@@ -1,7 +1,7 @@
 from coldtype.helpers import loopidx, interp_dict
-from coldtype.animation.timeable import Timing, Timeable, TimeableSet
-from coldtype.animation.timeline import Timeline
-from coldtype.animation.loop import Loop, LoopPhase
+from coldtype.time.timeable import Timing, Timeable, TimeableSet
+from coldtype.time.timeline import Timeline
+from coldtype.time.loop import Loop, LoopPhase
 
 class Frame():
     """
