+++ coldtype/time/easing.py
