+++ docs/api/reference/coldtype.animation.easing.ease.rst
@@ -1,6 +1,6 @@
-﻿coldtype.animation.easing.ease
+﻿coldtype.time.easing.ease
 ==============================
 
-.. currentmodule:: coldtype.animation.easing
+.. currentmodule:: coldtype.time.easing
 
 .. autofunction:: ease
