+++ examples/animations/808.md
@@ -8,7 +8,7 @@ For this animation, we'll need the standard coldtype library as well as some opt
 
 ```python
 from coldtype import *
-from coldtype.animation.midi import MidiReader
+from coldtype.time.midi import MidiReader
 from coldtype.warping import warp_fn
 ```
