+++ test/lyric_data.json
@@ -13,12 +13,12 @@
         {
           "text": "+llo",
           "start": 9,
-          "end": 20,
+          "end": 21,
           "idx": 1
         },
         {
           "text": "≈Worl",
-          "start": 20,
+          "start": 21,
           "end": 32,
           "idx": 2,
           "tidx": 0
